Split Vuex store into modules
